requirements-dev.txt
# ── Development & CI dependencies (ne pas installer en production) ───────────
-r requirements.txt

black==24.8.0
cfgv==3.4.0
coverage==7.6.1
distlib==0.4.0
filelock==3.20.0
flake8==7.1.1
identify==2.6.15
iniconfig==2.3.0
isort==5.13.2
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre-commit==3.8.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.2
pytest-django==4.8.0
virtualenv==20.35.3
wheel==0.45.1
